Move EXPECT_EXCEPTION macro to common utils header.
